Split Maple county stuff from cogs/ERLC.py into its own file cogs/MC.py
